% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runAlgoOnInstance}
\alias{runAlgoOnInstance}
\title{Runs algorithm on an instance with given parameters and returns performance.}
\usage{
runAlgoOnInstance(scen, instance, vals, seed = 123L, cutoff.time,
  show.info = TRUE)
}
\arguments{
\item{scen}{[\code{AClibScenario}]\cr
Scenario.}

\item{instance}{[\code{character(1)}]\cr
Name of instance to run on, must be from training or test set of \code{scen}.}

\item{vals}{[named \code{list}]\cr
Setting for parameters. Names are parameters names, elements their values.
If a parameter ist not listed here, its default value is taken from \code{scen$par.set}.
Missing values are removed last, before passing them to the algorithm script.}

\item{cutoff.time}{[\code{numeric(1)}]\cr
Cutoff time for algorithm run in seconds.
An upper limit is \code{scen$cutoff.time}.
Default is \code{scen$cutoff.time}.}
}
\description{
Runs algorithm on an instance with given parameters and returns performance.
}

